+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added back crafting recipes (raoulvdberge)
 - Changed Grid recipe (raoulvdberge)
 - Changed Crafting Monitor recipe (raoulvdberge)
+- Fixed Filters not persisting correctly in Wireless Grid and Wireless Crafting Monitor (raoulvdberge)
 
 ### 1.5
 - Port to Minecraft 1.12 (raoulvdberge)